Give incorrect box's ::before content a z-index of 1 to make it appear above the mine image again
This is needed because giving the mine image a filter gives it its own stacking context, which can make it appear on top of its sibling elements. Setting the z-index of the affected sibling element, in this case the ::before pseudo-element, will adjust it so that it will appear in front of the mine image once more.